When adjusting track volume/pan from Kontrol, the display is now updated correctly.
Even though CSurf_OnVolumeChange seems to adjust the track volume in REAPER, it seems we need to call CSurf_SetSurfaceVolume as well.
Otherwise, surfaces (including Kontrol) don't get notified about the change.
The same is true for the pan functions.
We originally did this, but this was regressed in 938e6ddc.