Tell Kontrol that the clip navigation controls are enabled.
Aside from ulluminating the lights, this also prevents NI Accessibility Helper from reporting "disabled" when using these controls to navigate to project start/end, markers, etc.
For now, these are always treated as enabled.
Strictly speaking, previous should be disabled when not at the start of the project, and next should be disabled when not at the end of the project/time selection/loop area.
However, this would involve polling the cursor position, which isn't ideal. For now, just always enabl.
Fixes #44.